move utils out of cmd and expose utils
helps things work without installing click
