+++ src/grafanarmadillo/cmd.py
@@ -1,7 +1,6 @@
 """Ready-to-run commands for common Grafana templating scenarios."""
 import json
-from pathlib import Path
-from typing import IO, Dict, NewType
+from typing import IO
 
 import click
 from grafana_client import GrafanaApi
@@ -9,12 +8,9 @@ from grafana_client import GrafanaApi
 from grafanarmadillo.alerter import Alerter
 from grafanarmadillo.dashboarder import Dashboarder
 from grafanarmadillo.find import Finder
-from grafanarmadillo.templator import Templator, findreplace
+from grafanarmadillo.templator import Templator, make_mapping_templator
+from grafanarmadillo.util import load_data
 
-
-EnvMapping = NewType("EnvMapping", Dict[str, Dict[str, str]])
-
-TOK_AUTO_MAPPING = "$auto"
 
 load_file_help = """Should be encoded as json. You can pass this in as a string; or as file using 'file://path/to/file'"""
 auto_template_env_help = "The special value '$auto' will automatically provide a value by prepending a '$' to the keys of the grafana mapping"
@@ -27,36 +23,6 @@ mapping_help = \
 env_grafana_help = "Name of the environment in the mapping file for Grafana, found in the `--mapping` argument"
 env_template_help = "Name of the environment in the mapping file for the template, found in the `--mapping` argument. " + \
 	auto_template_env_help
-
-
-def load_data(data_str: str):
-	"""Attempt to load data."""
-	_file_uri_prefix = "file://"
-	if data_str.startswith(_file_uri_prefix):
-		filename = Path(data_str.split(_file_uri_prefix)[1])
-		with filename.open(mode="r", encoding="utf-8") as data_file:
-			return json.load(data_file)
-	else:
-		return json.loads(data_str)
-
-
-def make_mapping_templator(mapping: EnvMapping, env_grafana: str, env_template: str) -> Templator:
-	"""Assemble the templator from the environment mapping."""
-	mapping_grafana = mapping[env_grafana]
-	if env_template == TOK_AUTO_MAPPING:
-		mapping_template = {k: "${%s}" % k for k in mapping_grafana.keys()}
-	else:
-		mapping_template = mapping[env_template]
-
-	# if some keys in the src mapping are not in the dst mapping
-	missing = mapping_grafana.keys() - mapping_template.keys()
-	if missing:
-		raise ValueError(f"Some keys in the source mapping are not present in the destination mapping. {missing=}")
-
-	grafana_to_template = {v: mapping_template[k] for k, v in mapping_grafana.items()}
-	template_to_grafana = {v: k for k, v in grafana_to_template.items()}
-
-	return Templator(make_template=findreplace(grafana_to_template), fill_template=findreplace(template_to_grafana))
 
 
 def make_grafana(config) -> GrafanaApi:
@@ -193,15 +159,6 @@ def import_alert(gfn: GrafanaApi, src: IO, dst: str, templator: Templator):
 	alert_info, folder_info = finder.create_or_get_alert(dst)
 	alert = templator.make_dashboard_from_template(alert_info, template)
 	alerter.import_alert(alert, folder_info)
-
-
-def resolve_object_to_filepath(base_path: Path, name: str):
-	"""Transform the "/folder/object" format to the path on disk that contains the template."""
-	path = Path(name)
-	if path.is_absolute():
-		path = path.relative_to("/")
-	template_path = (base_path / path).with_suffix(".json")
-	return template_path
 
 
 if __name__ == "__main__":
